src/requirements.txt
groq>=0.9.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
tavily-python>=0.3.0
supabase>=2.0.0
pdfplumber>=0.10.0
python-multipart>=0.0.9

fastapi
uvicorn
sse-starlette
